refactoring lots of vulkan stuff to support sprite pipeline alongside shape pipeline

+++ bindings/wrapper/NumpyWrapper.cpp
@@ -6,18 +6,18 @@ namespace griddy {
 template <class Scalar>
 class NumpyWrapper {
  public:
-  NumpyWrapper(std::vector<size_t> shape, std::vector<size_t> strides, std::unique_ptr<Scalar[]> data)
+  NumpyWrapper(std::vector<uint> shape, std::vector<uint> strides, std::unique_ptr<Scalar[]> data)
       : shape_(shape), strides_(strides), data_(std::move(data)) {
   }
 
-  std::vector<size_t> getShape() const { return shape_; }
-  std::vector<size_t> getStrides() const { return strides_; }
-  size_t getScalarSize() const { return sizeof(Scalar); }
+  std::vector<uint> getShape() const { return shape_; }
+  std::vector<uint> getStrides() const { return strides_; }
+  uint getScalarSize() const { return sizeof(Scalar); }
   Scalar* getData() const { return data_.get(); }
 
  private:
-  const std::vector<size_t> shape_;
-  const std::vector<size_t> strides_;
+  const std::vector<uint> shape_;
+  const std::vector<uint> strides_;
   const std::unique_ptr<Scalar[]> data_;
 };
 }  // namespace griddy
