+++ src/Griddy/Core/Observers/Observer.hpp
@@ -15,8 +15,8 @@ class Observer {
    */
   virtual std::unique_ptr<uint8_t[]> observe(int playerId) = 0;
 
-  virtual std::vector<size_t> getShape() const = 0;
-  virtual std::vector<size_t> getStrides() const = 0;
+  virtual std::vector<uint> getShape() const = 0;
+  virtual std::vector<uint> getStrides() const = 0;
 
   virtual void init(uint gridWidth, uint gridHeight);
