+++ src/Griddy/Core/Observers/TileObserver.cpp
@@ -8,11 +8,11 @@ TileObserver::TileObserver(std::shared_ptr<Grid> grid): Observer(grid) {}
 
 TileObserver::~TileObserver() {}
 
-std::vector<size_t> TileObserver::getShape() const {
+std::vector<uint> TileObserver::getShape() const {
   return {1, grid_->getWidth(), grid_->getHeight()};
 }
 
-std::vector<size_t> TileObserver::getStrides() const {
+std::vector<uint> TileObserver::getStrides() const {
   return {1, 1, 10};
 }
